requirements.txt
pandas
tqdm
ipykernel
torch
transformers
pytorch-lightning
tensorboard
python-dotenv
nltk
scikit-learn
wandb
fastcoref